requirements.txt
neo4j
python-dotenv==1.0.0
openai==1.57.0
langchain
langchain-neo4j
flask==3.0.0
flask-cors==4.0.0
httpx==0.27.2
